wishListHook added
